+++ djpcms/templatetags/tablesorter.py
@@ -5,10 +5,10 @@ from djpcms.core.models import nicerepr
 register = template.Library()
 
 @register.inclusion_tag('djpcms/tablesorter.html')
-def totable(headers, queryset, djp = None, appmodel = None):
-    return table(headers, queryset, djp, appmodel)
+def totable(headers, queryset, djp = None, appmodel = None, nd = 3):
+    return table(headers, queryset, djp, appmodel, nd)
 
 @register.filter
-def nicevalue(value):
-    return nicerepr(value)
+def nicevalue(value, nd = 3):
+    return nicerepr(value, nd)