+++ CHANGELOG.rst
@@ -1,6 +1,9 @@
 Version 0.8.6 - Development
 =======================================
-* Added :mod:``djpcms.contrib.social`` application with ``AOuth``.
+* Injecting the instance or ``None`` of current ``url`` into ``request`` object.
+* Added more docs on views and applications.
+* Added :mod:``djpcms.contrib.social`` application with ``OAuth``. Still in ``alpha``.
+* **42 unit tests**.
 
 Version 0.8.5 (2010-Nov-16)
 =======================================